update api and service locator, update also model

+++ app/src/main/java/com/sbaiardi/holdmybeer/model/Amount.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class Amount (
 
-	@SerializedName("value") val value : Int,
+	@SerializedName("value") val value : Double,
 	@SerializedName("unit") val unit : String
 )
