+++ app/src/main/java/com/sbaiardi/holdmybeer/model/Boil_volume.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 
 data class Boil_volume (
 
-	@SerializedName("value") val value : Int,
+	@SerializedName("value") val value : Double,
 	@SerializedName("unit") val unit : String
 )
